# Efficient MXNet sampling in the multinomial distribution