@pengzhao-intel 2019-06-21
@reminisce @wkcn for review.